refactor(web): extract CoinGecko id resolution into a shared module
The coins/list index, its slug rules and the asset-to-id resolver were
private to the curated fallback. The price chart needs the same resolution,
so they move to their own module rather than being duplicated.

The slug rules in particular must not be copied: they have to match the
asset registry exactly (`&` becomes `and`, so "S&P Global" is
`sandp-global`), and a second copy is a second chance to drift.
