+++ apps/web/src/lib/market-data/curated-fallback.ts
@@ -4,6 +4,7 @@ import { getVariantByMint } from '@tokens/asset-registry';
 import { CURATED_LIST_ORDER, getCuratedTokenList, type CuratedTokenListId } from '@tokens/asset-registry/compat';
 import type { Token } from '@/lib/types';
 import { withTtl } from './cache';
+import { COINGECKO_BASE_URL, loadCoinGeckoIndex, resolveCoinGeckoId, slugify } from './coingecko-index';
 import { fetchTokenLiquidity } from './token-liquidity';
 import { runScannerQuery } from './tradingview';
 import { numberOrNull, stringOrNull } from './types';
@@ -103,19 +104,6 @@ const METAL_TICKERS: Record<string, string> = {
     copper: 'COMEX:HG1!',
 };
 
-/**
- * Must stay identical to the asset registry's slug rules — notably `&` becomes
- * `and`, so "S&P Global" is `sandp-global`. Diverging here silently breaks every
- * ampersand name (AT&T, Procter & Gamble, Hims & Hers).
- */
-function slugify(value: string): string {
-    return value
-        .trim()
-        .toLowerCase()
-        .replaceAll('&', 'and')
-        .replace(/[^a-z0-9]+/g, '-')
-        .replace(/^-+|-+$/g, '');
-}
 
 // ---------------------------------------------------------------- TradingView
 
@@ -231,42 +219,6 @@ async function loadMetalQuotes(): Promise<Map<string, Quote>> {
 
 // ------------------------------------------------------------------ CoinGecko
 
-const COINGECKO_BASE_URL = 'https://api.coingecko.com/api/v3';
-
-interface CoinGeckoIndex {
-    ids: Set<string>;
-    bySymbol: Map<string, string>;
-    byName: Map<string, string>;
-}
-
-const loadCoinGeckoIndex = withTtl(12 * 60 * 60_000, async (): Promise<CoinGeckoIndex> => {
-    const res = await fetch(`${COINGECKO_BASE_URL}/coins/list`, {
-        headers: { accept: 'application/json' },
-        signal: AbortSignal.timeout(30_000),
-        cache: 'no-store',
-    });
-    if (!res.ok) throw new Error(`CoinGecko coins/list HTTP ${res.status}`);
-
-    const list = (await res.json()) as { id?: unknown; symbol?: unknown; name?: unknown }[];
-    const ids = new Set<string>();
-    const bySymbol = new Map<string, string>();
-    const byName = new Map<string, string>();
-
-    for (const coin of list) {
-        const id = stringOrNull(coin.id);
-        if (!id) continue;
-        ids.add(id);
-        const symbol = stringOrNull(coin.symbol)?.toUpperCase();
-        if (symbol && !bySymbol.has(symbol)) bySymbol.set(symbol, id);
-        const name = stringOrNull(coin.name);
-        if (name) {
-            const key = slugify(name);
-            if (key && !byName.has(key)) byName.set(key, id);
-        }
-    }
-    return { ids, bySymbol, byName };
-});
-
 async function fetchCoinGeckoQuotes(ids: readonly string[]): Promise<Map<string, Quote>> {
     const quotes = new Map<string, Quote>();
     if (ids.length === 0) return quotes;
@@ -373,22 +325,6 @@ function resolveCuratedAssets(listId: CuratedTokenListId): CuratedAsset[] {
     }
 
     return assets;
-}
-
-function resolveCoinGeckoId(asset: CuratedAsset, index: CoinGeckoIndex): string | null {
-    if (asset.coingeckoId && index.ids.has(asset.coingeckoId)) return asset.coingeckoId;
-
-    const symbol = asset.symbol.toUpperCase();
-    if (symbol && index.bySymbol.has(symbol)) return index.bySymbol.get(symbol) ?? null;
-
-    for (const alias of asset.aliases) {
-        const key = slugify(alias);
-        if (index.ids.has(key)) return key;
-        const byName = index.byName.get(key);
-        if (byName) return byName;
-    }
-
-    return index.byName.get(asset.assetId) ?? null;
 }
 
 function resolveScreenerQuote(asset: CuratedAsset, index: ScreenerIndex): Quote | null {
